pyproject.toml
[project]
name = "pxq"
version = "0.1.0"
description = "A pueue-like CLI for local and RunPod job management"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Takeru", email = "takeru1205@users.noreply.github.com"}
]
keywords = ["cli", "job-queue", "runpod", "gpu", "task-queue"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Distributed Computing",
]
dependencies = [
    "typer>=0.9.0",
    "fastapi>=0.100.0",
    "httpx>=0.24.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.1.0",
    "uvicorn>=0.23.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/takeru1205/pxq"
Repository = "https://github.com/takeru1205/pxq.git"
Issues = "https://github.com/takeru1205/pxq/issues"

[project.scripts]
pxq = "pxq.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pxq"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.15.2",
]
